requirements.txt
fastapi
uvicorn
requests
python-telegram-bot
pydantic
pillow
websocket-client